🔄 Rename sctfcoil module to superconducting_tf_coil for consistency
