+++ process/main.py
@@ -88,9 +88,9 @@ from process.power import Power
 from process.pulse import Pulse
 from process.resistive_tf_coil import ResistiveTFCoil
 from process.scan import Scan
-from process.sctfcoil import SuperconductingTFCoil
 from process.stellarator import Neoclassics, Stellarator
 from process.structure import Structure
+from process.superconducting_tf_coil import SuperconductingTFCoil
 from process.tf_coil import TFCoil
 from process.utilities.f2py_string_patch import string_to_f2py_compatible
 from process.vacuum import Vacuum
