+++ process/superconducting_tf_coil.py
